requirement.txt
fastapi
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
'pydantic[email]'
